requirements.txt
loguru==0.7.2
PyYAML==6.0.1
pillow==10.1.0
tqdm==4.66.4
pybind11==2.11
click==8.1.7
numpy==1.24.1
wandb==0.16.6
torchvision==0.17.2
compressai==1.2.6